--- public/pages/Configuration/index.js
@@ -1,8 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import Configuration from './Configuration';
-
-export default Configuration;
