--- public/pages/QueryDetails/index.js
@@ -1,8 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import QueryDetails from './QueryDetails';
-
-export default QueryDetails;
